refactor(server): split workspace lifecycle store
